improved the code structure and improved the message searching functionality
