+++ next.config.js
@@ -1,11 +1,10 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  reactStrictMode: true,
   swcMinify: true,
   images: {
     domains: ['res.cloudinary.com'],
   },
-}
+};
 
 module.exports = {
   async rewrites() {
@@ -14,8 +13,8 @@ module.exports = {
         source: '/api/:path*',
         destination: 'https://api.example.com/:path*',
       },
-    ]
+    ];
   },
-}
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
