+++ package.json
@@ -18,6 +18,7 @@
     "mongoose": "^6.5.2",
     "multer": "^1.4.5-lts.1",
     "next": "^13.3.0",
+    "next-auth": "^4.22.1",
     "next-connect": "^0.13.0",
     "react": "^18.2.0",
     "react-bootstrap": "^1.5.2",
