+++ src/component/css/addMessage.module.scss
@@ -1,0 +1,68 @@
+.s {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background: #e3e5e7;
+
+  position: relative;
+  z-index: 5;
+
+  &__innerContainer {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    margin: 10px 0;
+  }
+
+  &__replyerContainer {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin: 10px 0;
+    padding: 0 50px;
+  }
+
+  &__replytoInnerContainer {
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+
+    background: #ced3d987;
+    margin-right: 10px;
+    padding: 5px 20px;
+    border-left: 10px solid #43d3d3;
+    border-radius: 10px;
+  }
+
+  &__nameTxt {
+    font-size: 14px;
+    font-weight: 800;
+  }
+
+  &__nameTxtOtherUser {
+    color: cyan;
+  }
+
+  &__messageTxt {
+    font-size: 14px;
+  }
+
+  &__crossIcon {
+    cursor: pointer;
+  }
+
+  &__input {
+    width: 100%;
+    height: 40px;
+    outline: none;
+    border: none;
+    border-radius: 8px;
+    margin: 0 1rem;
+    padding-left: 1rem;
+  }
+  &__svg {
+    cursor: pointer;
+    margin: 0 1rem;
+  }
+}
