+++ src/components/css/mainScreen.module.scss
@@ -5,11 +5,3 @@
   display: flex;
   width: 100%;
 }
-
-.s {
-  width: 450px;
-  border-right: 1px solid #acb2b8;
-}
-.s2 {
-  width: 100%;
-}
