--- src/components/rightSection/css/chatlist.module.scss
@@ -1,135 +1,0 @@
-.s {
-  width: 100%;
-
-  &__containerOuter {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    height: 100vh;
-  }
-
-  &__containerWithoutReply {
-    height: calc(100vh - 120px);
-  }
-
-  &__containerWithReplyOn {
-    height: calc(100vh - 160px);
-  }
-
-  &__headerUserName {
-    font-family: 'lato';
-    font-size: 18px;
-    font-weight: 600;
-    letter-spacing: 2px;
-
-    margin-left: 10px;
-    margin-bottom: 0;
-  }
-
-  &__headerContainer {
-    height: 60px;
-
-    position: relative;
-    background: #e3e5e7;
-
-    display: flex;
-    align-items: center;
-
-    padding-left: 30px;
-    z-index: 5;
-  }
-
-  &__chatContainer {
-    position: relative;
-    height: 100%;
-    overflow-y: auto;
-
-    &__overlay {
-      position: fixed;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-      background: url('https://cdna.artstation.com/p/assets/images/images/011/168/054/large/joel-morales-escobar-trama-j02.jpg?1528192578');
-      background-size: cover;
-      z-index: -1;
-      opacity: 0.5;
-    }
-  }
-
-  &__addAChatContainer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    background: #f0f2f5;
-
-    position: relative;
-    z-index: 5;
-
-    &__innerContainer {
-      display: flex;
-      align-items: center;
-      width: 100%;
-      margin: 10px 0;
-    }
-
-    &__replyerContainer {
-      width: 100%;
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      margin: 10px 0;
-      padding: 0 50px;
-    }
-
-    &__replytoInnerContainer {
-      width: 100%;
-      display: flex;
-      flex-direction: column;
-
-      background: #ced3d987;
-      margin-right: 10px;
-      padding: 5px 20px;
-      border-left: 10px solid #43d3d3;
-      border-radius: 10px;
-    }
-
-    &__nameTxt {
-      font-size: 14px;
-      font-weight: 800;
-    }
-
-    &__nameTxtOtherUser {
-      color: cyan;
-    }
-
-    &__messageTxt {
-      font-size: 14px;
-    }
-
-    &__crossIcon {
-      cursor: pointer;
-    }
-
-    &__input {
-      width: 100%;
-      height: 40px;
-      outline: none;
-      border: none;
-      border-radius: 8px;
-      margin: 0 1rem;
-      padding-left: 1rem;
-    }
-    &__svg {
-      cursor: pointer;
-      margin: 0 1rem;
-    }
-  }
-
-  &__noUserSelectedContainer {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, calc(-50% - 60px));
-  }
-}
