+++ src/components/userList/css/userList.module.scss
@@ -1,4 +1,6 @@
 .s1 {
+  width: 450px;
+  border-right: 1px solid #acb2b8;
   background-color: white;
 
   &__headerContainer {
